Refine interface of image decoder adapter

+++ ohos_interface/ohos_glue/ohos_adapter/bridge/webcore/ark_ohos_image_decoder_adapter_wrapper.cpp
@@ -49,7 +49,7 @@ bool ArkOhosImageDecoderAdapterWrapper::DecodeToPixelMap(const uint8_t* data,
                                                          OHOS::NWeb::AllocatorType type,
                                                          bool useYuv)
 {
-    return ctocpp_->DecodeToPixelMap(data, size, type, useYuv);
+    return ctocpp_->DecodeToPixelMap(data, size, (uint32_t)type, useYuv);
 }
 
 int32_t ArkOhosImageDecoderAdapterWrapper::GetFd()
